maritime-rl/requirements.txt
numpy>=1.22.0
python-dotenv>=1.0.0
httpx>=0.24.0
attrs>=23.0.0
cachetools>=5.0.0
fastavro>=1.7.0
confluent-kafka>=2.2.0
requests>=2.27.0
pydantic>=2.0.0
authlib>=1.5.0 